Patching frigate reward crash workaround
Swapping the overall expedition reward with the Phase 5 reward, and make Phase 5 optional to avoid being stuck in the expedition once it is completed. You cannot claim the Organic Frigate (it will still be unlocked in the QuickSilver vendor account-wide), but you can end the expedition at the Anomaly expedition terminal.